+++ Src/XmlDocInspections.Plugin/Settings/XmlDocInspectionsOptionsPage.cs
@@ -4,9 +4,16 @@ using System.Linq.Expressions;
 using JetBrains.Application.Settings;
 using JetBrains.DataFlow;
 using JetBrains.ReSharper.Feature.Services.Daemon.OptionPages;
+#if RESHARPER20163 || RESHARPER20171
 using JetBrains.UI.Options;
 using JetBrains.UI.Options.OptionsDialog2.SimpleOptions;
 using JetBrains.UI.Resources;
+#else
+using JetBrains.Application.UI.Icons.CommonThemedIcons;
+using JetBrains.Application.UI.Options;
+using JetBrains.Application.UI.Options.OptionsDialog;
+
+#endif
 
 namespace XmlDocInspections.Plugin.Settings
 {