+++ History.md
@@ -1,3 +1,6 @@
+### vNext ###
+- ReSharper 2017.2 support
+
 ### 1.8.0 ###
 - Added a quick fix action on XML Doc warnings to generate doc templates
 - Added a configuration option to include/exclude constructors