+++ History.md
@@ -1,4 +1,4 @@
-### vNext ###
+### 2.0.0 ###
 - Improved options page (added clear distinction between inclusion and exclusion rules) => this means a breaking change for "Exclude constructors" and "Exclude members which override super/base members" settings
 - Add attributes inclusion setting (configured by default to `[PublicAPI]`)
 - ReSharper 2017.2 support